requirements.txt
asgiref==3.8.1
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.6
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.2
factory-boy==3.3.0
Faker==25.9.1
idna==3.7
iniconfig==2.0.0
mysqlclient==2.2.4
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pluggy==1.5.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
reportlab==4.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
XlsxWriter==3.2.0
